% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-wordcloud.R
\name{geom_text_wordcloud}
\alias{geom_text_wordcloud}
\title{Wordcloud text geom.}
\description{
\code{geom_text_wordcloud} adds text to the plot using a variation of the wordcloud2.js algorithm.
}
\details{
This geom is based on \code{\link[ggrepel]{geom_text_repel}} which in turn is
based on \code{\link[ggplot2]{geom_text}}. See the documentation for those
functions for more details.
}
