% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwordcloud-wrapper.R
\name{ggwordcloud2}
\alias{ggwordcloud2}
\title{\code{ggwordcloud2} is an approximate replacement for \code{\link[wordcloud2]{wordcloud2}}. It has almost the same syntax but fewer options. In particular, there is no background image (so far...).}
\usage{
ggwordcloud2(data, minRotation = -pi/4, maxRotation = pi/4,
  shuffle = TRUE, rotateRatio = 0.4, ellipticity = 0.65, ...)
}
\arguments{
\item{data}{a dataframe whose two first columns are the names and the freqs or a table}

\item{minRotation}{the minimal rotation angle}

\item{maxRotation}{the maximal rotation angle}

\item{shuffle}{if TRUE, the words are shuffled at the beginning}

\item{ellipticity}{control the eccentricity of the wordcloud}

\item{...}{the remaining parameters are passed to geom_text_wordcloud}

\item{rotateRation}{the proportion of rotated words}
}
\value{
a ggplot
}
\description{
\code{ggwordcloud2} is an approximate replacement for \code{\link[wordcloud2]{wordcloud2}}. It has almost the same syntax but fewer options. In particular, there is no background image (so far...).
}
