% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/love_words.R
\docType{data}
\name{love_words}
\alias{love_words}
\alias{love_words_small}
\title{Love in several languages with number of speakers}
\format{a data.frame with 147 observations (or 34 for the small one) of 4
variables
\describe{
\item{lang}{the ISO 639 language code}
\item{word}{the word love in that language}
\item{native_speakers}{number of native speakers in millions}
\item{speakers}{number of speakers in millions}
}}
\source{
wikipedia
}
\usage{
love_words

love_words_small
}
\description{
A dataset containing the word love in different languages (147 or 34 for the
small one) as well as the number of native speakers and overall speakers of
those languages.
}
\keyword{datasets}
