% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_trans.R
\name{power_trans}
\alias{power_trans}
\title{A signed power transform}
\usage{
power_trans(power = 1)
}
\arguments{
\item{power}{power exponent of the direct transform}
}
\description{
A signed power transform
}
