% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thankyou_words.R
\docType{data}
\name{thankyou_words}
\alias{thankyou_words}
\alias{thankyou_words_small}
\title{'Thank you' in several languages with number of speakers}
\format{a data.frame with 133 observations (or 34 for the small one) of 4
variables
\describe{
\item{iso_639_3}{the ISO 639-3 language code}
\item{word}{the word love in that language}
\item{native_speakers}{number of native speakers in millions}
\item{speakers}{number of speakers in millions}
}}
\source{
wikipedia
}
\usage{
thankyou_words

thankyou_words_small
}
\description{
A dataset containing the word 'Thank you' in different languages (133 or 34 for the
small one) as well as the number of native speakers and overall speakers of
those languages.
}
\keyword{datasets}
